requirements.txt
tensorflow==2.10.0
flask==2.0.1
numpy==1.23.5
pandas==1.5.3
pillow==9.3.0
matplotlib==3.6.3
werkzeug==2.0.1
scikit-learn==1.2.2